refactor(explain): extract filter_strategy module from plan_builder (#471)
plan_builder.rs was 501 NLOC — 1 line over the CONTRIBUTING.md 500 NLOC
file limit (Devin Finding G on PR #606). Extract the filter-strategy
decision logic into a dedicated sub-module:

- New crates/velesdb-core/src/velesql/explain/filter_strategy.rs (112 NLOC)
  exports pub(super): resolve_filter_strategy, strip_vector_predicates,
  estimate_filter_stats + the three FALLBACK_SELECTIVITY_THRESHOLD /
  PREFILTER_RECALL_GUARD / POSTFILTER_TOPK_COST_FRACTION constants.

- plan_builder.rs shrinks to 386 NLOC and delegates to the new module
  via free functions (no public API change, logic bit-for-bit identical).

The new module also carries a dedicated doc section (Finding I,
non-blocker analysis) explaining why PreFilter is rarely chosen on
large collections by design, and points to follow-up issue #609 for
the k*cpu_tuple_cost post-filter refinement.

Tests: all 7 explain_cost_calibrated BDD tests still pass.
