+++ crates/velesdb-core/src/velesql/explain/plan_builder.rs
@@ -5,6 +5,7 @@
 
 use std::collections::HashSet;
 
+use super::filter_strategy::{estimate_filter_stats, resolve_filter_strategy};
 use super::formatter;
 use super::node_stats;
 use super::types::{
@@ -16,32 +17,7 @@ use crate::collection::search::query::match_planner::{
 };
 use crate::collection::stats::CollectionStats as CoreCollectionStats;
 use crate::velesql::ast::{Condition, LetBinding, SelectStatement};
-use crate::velesql::cost_estimator::CostEstimator;
 use crate::velesql::MatchClause;
-
-/// Threshold (selectivity) above which the fallback heuristic switches from
-/// PreFilter to PostFilter. Preserved unchanged for backward compatibility
-/// with tests that predate calibrated costs (see
-/// `test_filter_strategy_post_filter_default`).
-const FALLBACK_SELECTIVITY_THRESHOLD: f64 = 0.1;
-
-/// Guardrail: if a predicate filters out less than this fraction of rows,
-/// running the pre-filter before HNSW is likely to hurt recall (too many
-/// candidates survive). Forces PostFilter in that case.
-const PREFILTER_RECALL_GUARD: f64 = 0.5;
-
-/// Approximate fraction of the filter-scan cost attributed to the
-/// post-filter step. Post-filter only evaluates the predicate on the
-/// top-k HNSW results (typically k = 10) while the filter-scan formula
-/// scales with `total * selectivity`. Using a fixed `0.01` multiplier is
-/// a rough but conservative approximation — it slightly overestimates
-/// the post-filter cost for very large collections with high selectivity
-/// (Devin finding on PR #606), yet the HNSW term dominates in those
-/// regimes so it does not flip strategy decisions in practice. A future
-/// refinement could replace this with a proper `k * cpu_tuple_cost`
-/// model once `k` and the calibrated `cpu_tuple_cost` are threaded
-/// through to `resolve_filter_strategy`.
-const POSTFILTER_TOPK_COST_FRACTION: f64 = 0.01;
 
 impl QueryPlan {
     /// Creates a new query plan from a SELECT statement.
@@ -312,8 +288,9 @@ impl QueryPlan {
         let mut filter_strategy = FilterStrategy::None;
 
         if !filter_conditions.is_empty() {
+            let heuristic_fallback = Self::estimate_selectivity(filter_conditions);
             let (selectivity, estimation_method, estimated_rows) =
-                Self::estimate_filter_stats(stmt, filter_conditions, stats);
+                estimate_filter_stats(stmt, heuristic_fallback, stats);
 
             // Reason: plan_builder owns stmt so the real ef_search/candidates
             // are the same values used in `build_scan_node` above
@@ -323,7 +300,7 @@ impl QueryPlan {
             let ef_search = Self::resolve_ef_search(stmt);
             let candidates = u32::try_from(stmt.limit.unwrap_or(50)).unwrap_or(u32::MAX);
 
-            filter_strategy = Self::resolve_filter_strategy(
+            filter_strategy = resolve_filter_strategy(
                 selectivity,
                 has_vector_search,
                 ef_search,
@@ -347,191 +324,6 @@ impl QueryPlan {
         }
 
         filter_strategy
-    }
-
-    /// Computes `(selectivity, estimation_method, estimated_rows)` for the
-    /// filter block. When `stats` is `Some` and the WHERE clause tree is
-    /// available, uses `CostEstimator::estimate_condition_selectivity_with_method`
-    /// on the non-vector subtree and reports the actual method that produced
-    /// the estimate (histogram / cardinality / heuristic — issue #471, Devin
-    /// finding 2). Otherwise falls back to the string-count heuristic.
-    fn estimate_filter_stats(
-        stmt: &SelectStatement,
-        filter_conditions: &[String],
-        stats: Option<&CoreCollectionStats>,
-    ) -> (f64, Option<String>, Option<u64>) {
-        match (stats, stmt.where_clause.as_ref()) {
-            (Some(s), Some(where_cond)) => {
-                let est = CostEstimator::new(s);
-                let non_vector = Self::strip_vector_predicates(where_cond);
-                let (sel, method) = non_vector.as_ref().map_or(
-                    (
-                        1.0,
-                        crate::velesql::cost_estimator::SelectivityMethod::Heuristic,
-                    ),
-                    |c| est.estimate_condition_selectivity_with_method(c),
-                );
-                let total = s.total_points.max(s.row_count);
-                // Reason: total*sel is a cardinality estimate; small fractional
-                // losses after ceil are acceptable for EXPLAIN display.
-                #[allow(
-                    clippy::cast_precision_loss,
-                    clippy::cast_possible_truncation,
-                    clippy::cast_sign_loss
-                )]
-                let rows = ((total as f64) * sel).ceil() as u64;
-                (sel, Some(method.as_str().to_string()), Some(rows))
-            }
-            _ => (Self::estimate_selectivity(filter_conditions), None, None),
-        }
-    }
-
-    /// Returns the where-clause condition tree with `VectorSearch`-family
-    /// nodes removed, because selectivity of the non-vector predicates is
-    /// what drives the pre/post-filter decision.
-    ///
-    /// Returns `None` when every branch is vector-related (selectivity = 1.0).
-    fn strip_vector_predicates(condition: &Condition) -> Option<Condition> {
-        match condition {
-            Condition::VectorSearch(_)
-            | Condition::VectorFusedSearch(_)
-            | Condition::SparseVectorSearch(_)
-            | Condition::Similarity(_) => None,
-            Condition::And(left, right) => {
-                match (
-                    Self::strip_vector_predicates(left),
-                    Self::strip_vector_predicates(right),
-                ) {
-                    (Some(l), Some(r)) => Some(Condition::And(Box::new(l), Box::new(r))),
-                    (Some(l), None) => Some(l),
-                    (None, Some(r)) => Some(r),
-                    (None, None) => None,
-                }
-            }
-            Condition::Or(left, right) => {
-                match (
-                    Self::strip_vector_predicates(left),
-                    Self::strip_vector_predicates(right),
-                ) {
-                    (Some(l), Some(r)) => Some(Condition::Or(Box::new(l), Box::new(r))),
-                    // If one branch is vector-only, the OR is satisfied by
-                    // the vector side too → treat as selectivity 1.0.
-                    _ => None,
-                }
-            }
-            Condition::Not(inner) => {
-                Self::strip_vector_predicates(inner).map(|c| Condition::Not(Box::new(c)))
-            }
-            Condition::Group(inner) => {
-                Self::strip_vector_predicates(inner).map(|c| Condition::Group(Box::new(c)))
-            }
-            other => Some(other.clone()),
-        }
-    }
-
-    /// Picks `PreFilter` vs `PostFilter` given a pre-computed selectivity,
-    /// the real `ef_search` / `candidates` from the query, and optional
-    /// calibrated stats.
-    ///
-    /// When `stats` is `Some` and there's a vector search in the plan, runs
-    /// a cost comparison: pre-filter cost (scan + filter then HNSW on the
-    /// reduced set) vs post-filter cost (HNSW then filter on k results).
-    /// The HNSW cost uses `estimate_hnsw_search_cost_with_ef(ef_search,
-    /// candidates)` so the comparison reflects the user's actual WITH clause
-    /// (issue #471, Devin finding 4) instead of a hard-coded `k = 10`.
-    /// A recall guardrail forces `PostFilter` when selectivity is too loose
-    /// (>=0.5) so the pre-scan does not starve HNSW of good candidates.
-    ///
-    /// When `stats` is `None`, preserves the historical 0.1 threshold.
-    fn resolve_filter_strategy(
-        selectivity: f64,
-        has_vector_search: bool,
-        ef_search: u32,
-        candidates: u32,
-        stats: Option<&CoreCollectionStats>,
-    ) -> FilterStrategy {
-        let Some(s) = stats else {
-            return if selectivity > FALLBACK_SELECTIVITY_THRESHOLD {
-                FilterStrategy::PostFilter
-            } else {
-                FilterStrategy::PreFilter
-            };
-        };
-
-        if !has_vector_search {
-            // No vector search → the notion of pre/post-filter doesn't apply
-            // the same way; preserve the historical threshold for parity with
-            // the non-vector plan shape.
-            //
-            // Observation (Devin finding on PR #606): for non-vector queries,
-            // the threshold is applied to a calibrated histogram-based
-            // selectivity once `stats` are present, but to the
-            // `0.5^n`-style heuristic when `stats` is `None`. A highly
-            // selective predicate such as `price < 10` (sel ≈ 0.01 after
-            // ANALYZE vs. 0.5 without) therefore flips the reported
-            // `filter_strategy` from PostFilter (heuristic) to PreFilter
-            // (calibrated). This is purely informational for non-vector
-            // plans — the execution engine does not dispatch on this
-            // field when there is no HNSW stage — but is visible in
-            // EXPLAIN output. Kept intentionally: the calibrated answer
-            // is the semantically correct one.
-            return if selectivity > FALLBACK_SELECTIVITY_THRESHOLD {
-                FilterStrategy::PostFilter
-            } else {
-                FilterStrategy::PreFilter
-            };
-        }
-
-        // Recall guardrail: a loose filter leaves too many candidates in the
-        // HNSW frontier, so it's safer to post-filter after HNSW has done
-        // the heavy lifting on the full set.
-        if selectivity >= PREFILTER_RECALL_GUARD {
-            return FilterStrategy::PostFilter;
-        }
-
-        let est = CostEstimator::new(s);
-        let hnsw_cost = est
-            .estimate_hnsw_search_cost_with_ef(ef_search, candidates)
-            .total();
-        // Pre-filter: evaluate the predicate on **every** row of the
-        // collection (scan proportional to `total`, not `total*sel`) — then
-        // run HNSW on the `sel*total` surviving candidates. HNSW on a
-        // reduced set scales as `(ef + k) * log2(total*sel)`, not linearly
-        // in the reduction factor, hence the dedicated
-        // `estimate_hnsw_search_cost_with_ef_on_size` call (Devin finding E
-        // on #606). The filter-scan component uses selectivity=1.0 so the
-        // CBO does not under-estimate pre-filter cost by `1/selectivity`
-        // (Devin finding A on #606).
-        let total_points = s.total_points.max(s.row_count).max(1);
-        // Reason: `total_points * selectivity` is a cardinality estimate;
-        // floor to u64 is acceptable for a reduced-set size used in
-        // `log2(size)` probe counting.
-        #[allow(
-            clippy::cast_precision_loss,
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss
-        )]
-        let reduced_size = ((total_points as f64) * selectivity).max(1.0) as u64;
-        let hnsw_on_reduced = est
-            .estimate_hnsw_search_cost_with_ef_on_size(ef_search, candidates, reduced_size)
-            .total();
-        let pre_filter = est.estimate_filter_cost_from_selectivity(1.0).total() + hnsw_on_reduced;
-        // Post-filter: full HNSW pass, then filter evaluation on the
-        // top-k results (small constant cost, approximated by
-        // `POSTFILTER_TOPK_COST_FRACTION` applied to the filter cost on
-        // the reduced row-set — see the constant's doc-comment for the
-        // rationale and follow-up plan).
-        let post_filter = hnsw_cost
-            + est
-                .estimate_filter_cost_from_selectivity(selectivity)
-                .total()
-                * POSTFILTER_TOPK_COST_FRACTION;
-
-        if pre_filter < post_filter {
-            FilterStrategy::PreFilter
-        } else {
-            FilterStrategy::PostFilter
-        }
     }
 
     /// Analyzes a condition to extract vector search and filter info.
